interview/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
beanie==1.23.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6 